:arrow_up_small: 1.32 :arrow_up_small:

+++ Charlotte.py
@@ -6,6 +6,7 @@ import logging
 import traceback
 import requests
 import json
+import random
 from colorama import init
 from colorama import Fore, Back, Style
 from urllib.parse import urlencode
@@ -33,10 +34,14 @@ db.autocommit(True)
 db.ping(True)
 
 #Constants
-version = 1.27
-servers = ['yozora', 'ainu', 'kotorikku', 'kawata', 'toh.ru', 'ryusei', 'ryu-sei', 'waving', 'enjuu', 'verge', 'toh ru', 'toh-ru', 'katori']
-emailChecks = ['verify e', 'verification', 'on email', 'verify m', 'verify a', 'email t']
+version = 1.32
+filters = ['yozora', 'ainu', 'kotorikku', 'kawata', 'ryusei', 'ryu-sei', 'enjuu', 'verge', 'katori', 'nigger', 'discord.gg/', 'gatari', 'ripple'] # bad boy words
+emailChecks = ['verify e', 'verification', 'on email', 'verify m', 'verify a', 'email t', 'w verify', 'i verify']
 SQLChecks = [';', 'drop', 'ripple', 'select', '*'] # because im paranoid as fuck
+
+# a list of our psychedelic thoughts..
+cmyuiPsych = ['pentap AiAe [a bit jumpy] HD', 'i thought i was going to bite my tongue but it was my immune system', 'smoke', 'seeing everything even when i closed my eyes', 'pulsating teleportation', 'feeling like we had something to do', 'nose bleeding hallucination', 'shit horse vodka', 'd', 'lost job', 'vomit', 'chicken nuggets', 'beach', 'fat guy', 'glass ceiling', 'facial hair', 'waving green walls', 'sea land', 'rich', 'pleasure', 'reaching for exponential pleasure', 'self dissection', 'high (first lyric in song)', 'follow the light', 'paul', 'leo', 'francesco', 'dad offering beer', 'aero chord', 'acid interstate', 'driving through the jungles of germany', 'jaegarmeister in coke', 'wasp swarm', 'balcony', 'bending mushroom houses', 'panorama', 'dark green red and blue treetops', 'bugs in the forest', 'water dam', 'dark forest pathways', 'controlling your brain', 'begging for life and forgiveness', 'jacob crying', 'psychedelic music', 'akatsuki event', 'mirkas scarf', 'i am gonnnnnnnnneeeeeeeeeeee', 'hair splitting reality', 'hugging jacob', 'donkey kong fur flowing', 'calendar pulsating', 'flowers moving', 'clouds on the ceiling', 'horses and bodies on the design with lighting', 'pins and needles stabbing my fingers', 'drawing blood and becoming cold', 'going down the water slide', 'asking a girl for a key card', 'running into their room and bed', 'waking up in destroyed room with paramedics', 'becoming older and older and then god himself', 'hearing them say i was dead', 'the jungle under my monitor', 'cat and dog', 'dinner with shinis family', 'fett', 'wanchu back', 'leo\'s sweater', 'eating a sandwich', 'sitting on the bus', 'eating one chicken nugget']
+justicePsych = ['none yet']
 
 # Startup, after login action
 @client.event
@@ -84,30 +89,26 @@ async def on_message(message):
         print(Fore.CYAN + "Report recieved. It has been moved to #reports{end}".format(end=" and sent to {}.".format(config['discord']['username']) if config['default']['report_pm'] == 1 else "."))
     elif message.author != client.user:
         # Checks for things in message
-        if any(x in message.content.lower() for x in emailChecks):
-            if message.server is None:
-                if "badge" not in message.content.lower():
-                    await client.send_message(message.author, 'Right, this is an automated message as it was presumed your message was about: Email Verification\n\nAs the verification page says, Akatsuki does not use verification emails. To verify your account, simply install the switcher, install the certificate, click the server you\'d like to play on, and click On/Off, then login to osu! to complete the verification process.')
-                    if int(config['default']['debug']) == 1:
-                        print(Fore.MAGENTA + "Triggered: Verification Email Support\nUser: {}".format(message.author))
-                    else:
-                        print(Fore.MAGENTA + "Aborted Trigger: Email Verification Support, due to \"badge\" contents of the message.\nUser: {}".format(message.author))
-            # Akatsuki's ServerID
-            elif message.server.id == config['akatsuki']['server_id']:
-                if "badge" not in message.content.lower():
-                    await client.send_message(message.author, 'Right, this is an automated message as it was assumed you needed assitance in Akatsuki with: Email Verification\n\nAs the verification page says, Akatsuki does not use verification emails. To verify your account, simply install the switcher, install the certificate, click the server you\'d like to play on, and click On/Off, then login to osu! to complete the verification process.')
-                    await client.delete_message(message)
-                    if int(config['default']['debug']) == 1:
-                        print(Fore.MAGENTA + "Triggered: Verification Email Support\nUser: {}".format(message.author))
-                else:
-                    print(Fore.MAGENTA + "Aborted Trigger: Email Verification Support, due to \"badge\" contents of the message.\nUser: {}".format(message.author))
-        elif any(x in message.content.lower() for x in servers) or 'discord.gg/' in message.content.lower() and message.author.id != config['discord']['owner_id']:
+        if any(x in message.content.lower() for x in emailChecks) and message.server.id == config['akatsuki']['server_id']:
+            if "badge" not in message.content.lower():
+                await client.send_message(message.author, 'Right, this is an automated message as it was assumed you needed assitance in Akatsuki with: Email Verification\n\nAs the verification page says, Akatsuki does not use verification emails. To verify your account, simply install the switcher, install the certificate, click the server you\'d like to play on, and click On/Off, then login to osu! to complete the verification process.')
+                await client.delete_message(message)
+                if int(config['default']['debug']) == 1:
+                    print(Fore.MAGENTA + "Triggered: Verification Email Support\nUser: {}".format(message.author))
+            else:
+                print(Fore.MAGENTA + "Aborted Trigger: Email Verification Support, due to \"badge\" contents of the message.\nUser: {}".format(message.author))
+
+        elif any(x in message.content.lower() for x in filters) and message.author.id != config['discord']['owner_id']:
             await client.delete_message(message)  # Remove other private server-related messages
+            await client.send_message(message.author, 'Hello,\n\nYour message in osu!Akatsuki has been removed as it has been deemed unsuitable.\n```{}```'.format(message.content))
+            print(Fore.MAGENTA + "Filtered message | '{}: {}'".format(message.author, message.content))
 
         if message.server is None: # Private messages
                 print(Fore.YELLOW + Style.BRIGHT + "{} [{}] {}: {}".format(message.timestamp, message.channel, message.author, message.content))
+
         elif config['discord']['owner_id'] in message.content: # When you are pinged
                 print(Fore.CYAN + Style.BRIGHT + "{} [{} ({})] {}: {}".format(message.timestamp, message.server, message.channel, message.author, message.content))
+
         elif (config['discord']['username'] in message.content.lower() and len(config['discord']['username']) > 1): # When your username is mentioned (either actual one, or custom set in configuration)
                 print(Fore.GREEN + Style.BRIGHT + "{} [{} ({})] {}: {}".format(message.timestamp, message.server, message.channel, message.author, message.content))
 
@@ -195,6 +196,7 @@ async def on_message(message):
                     except:
                         await client.send_message(message.channel, 'something exploded. L')
 
+            """
                 elif messagecontent[0].lower() == '$partner':
                     userID = messagecontent[1]
                     streamName = messagecontent[2]
@@ -212,10 +214,6 @@ async def on_message(message):
                     cursor.execute("SELECT * FROM partners WHERE userid = {}".format(userID))
                     result = cursor.fetchone()
                     if result is None:
-                        """ If u ever figure out how to ping someone to give them a role. yep.
-                        role = discord.utils.get(message.server.roles, id=533300101972623371)
-                        await client.add_roles(message.author, role)
-                        """
                         cursor.execute("INSERT INTO partners (userid, stream_username, platform) VALUES ('{}', '{}', '{}')".format(userID, streamName, platform))
                         await client.send_message(message.channel, "{} has been sucessfully registered as an Akatsuki partner.".format(streamName))
                     else:
@@ -233,7 +231,7 @@ async def on_message(message):
                     else:
                         await client.send_message(message.channel, "{} is not a registered partner.".format(streamName))
 
-            """ otherwise
+            
             Process regular user command.
             """
             if messagecontent[0].lower() == '$user': # akatsuki userinfo command
@@ -276,6 +274,13 @@ async def on_message(message):
 
                 await client.send_message(message.channel, embed=embed)
                 await client.delete_message(processingMessage)
+
+            elif messagecontent[0].lower() == '$d': # d
+                cj = random.randint(1, 10)
+                if cj <= 7: # 70% chance cmyui, since his list bigger
+                    await client.send_message(message.channel, 'cmyui: {}'.format(random.choice(cmyuiPsych)))
+                else:
+                    await client.send_message(message.channel, 'Justice: {}'.format(random.choice(justicePsych)))
 
             elif messagecontent[0].lower() == '$faq': # FAQ command
                 try:
